+++ pgfinder/matching.py
@@ -4,7 +4,6 @@ from decimal import Decimal
 from pgfinder import MULTIMERS, MOD_TYPE, MASS_TO_CLEAN
 import pandas as pd
 
-from pgfinder.utils import calculate_ppm_delta
 from pgfinder.logs.logs import LOGGER_NAME
 
 LOGGER = logging.getLogger(LOGGER_NAME)
@@ -468,3 +467,41 @@ def data_analysis(
     cleaned_data_df.attrs["ppm"] = user_ppm
 
     return cleaned_data_df
+
+
+def calculate_ppm_delta(
+    df: pd.DataFrame,
+    observed: str = "mwMonoisotopic",
+    theoretical: str = "theo_mwMonoisotopic",
+    diff: str = "diff_ppm",
+) -> pd.DataFrame:
+    """Calculate the difference in Parts Per Million between observed and theoretical masses.
+
+    The PPM difference between observed and theoretical mass is calculated as...
+
+    .. math:: (1000000 * (obs - theor)) / theor
+
+    The function ensures the column is placed after the theoretical mass column to facilitate its use.
+
+    Parameters
+    ----------
+    df : pd.DataFrame
+        Pandas DataFrame of results.
+    observed : str
+        Variable that defines the observed PPM.
+    theoretical : str
+        Variable that defines the theoretical PPM.
+    diff: str
+        Variable to be created that holds the difference in PPM.
+
+    Returns
+    -------
+    pd.DataFrame
+        Pandas DataFrame with difference noted in column diff_ppm.
+
+    """
+    column_order = list(df.columns)
+    theoretical_position = column_order.index(theoretical) + 1
+    df.insert(theoretical_position, diff, (1000000 * (df[observed] - df[theoretical])) / df[theoretical])
+    LOGGER.info("Difference in PPM calculated.")
+    return df
