+++ tests/test_utils.py
@@ -2,10 +2,8 @@
 from decimal import Decimal
 from pathlib import Path
 
-import numpy as np
-import pandas as pd
 
-from pgfinder.utils import convert_path, update_config, dict_to_decimal, calculate_ppm_delta
+from pgfinder.utils import convert_path, update_config, dict_to_decimal
 
 
 def test_convert_path(tmpdir) -> None:
@@ -38,12 +36,3 @@ def test_dict_to_decimal() -> None:
     assert isinstance(decimal_dict["a"], Decimal)
     assert isinstance(decimal_dict["b"]["c"], Decimal)
     assert isinstance(decimal_dict["d"], str)
-
-
-def test_calculate_ppm_delta() -> None:
-    """Test addition of PPM column."""
-    SAMPLE_DF = pd.DataFrame({"obs": [5645.35435454, 1, 879546.3924093], "exp": [3954.49849514, 2, 879546.8974916]})
-    DELTA_DF = pd.DataFrame({"diff_ppm": [427577.82345296827, -500000.0, -0.5742528357381609]})
-    TARGET_DF = pd.concat([SAMPLE_DF, DELTA_DF], axis=1)
-
-    pd.testing.assert_frame_equal(calculate_ppm_delta(SAMPLE_DF, observed="obs", theoretical="exp"), TARGET_DF)
